# swagger.json
{"openapi":"3.0.0","paths":{"/api/health":{"get":{"operationId":"AppController_getHealth","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["App"]}},"/api/tasks":{"post":{"operationId":"TasksController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":"Task created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestError"}}}}},"tags":["Tasks"],"security":[{"Bearer":[]}]},"get":{"operationId":"TasksController_findMany","parameters":[],"responses":{"200":{"description":"Get tasks by params.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestError"}}}}},"tags":["Tasks"],"security":[{"Bearer":[]}]}},"/api/tasks/search":{"get":{"operationId":"TasksController_search","parameters":[],"responses":{"200":{"description":"Search by title or description.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestError"}}}}},"tags":["Tasks"],"security":[{"Bearer":[]}]}},"/api/tasks/{id}":{"get":{"operationId":"TasksController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get task by id.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestError"}}}}},"tags":["Tasks"],"security":[{"Bearer":[]}]},"put":{"operationId":"TasksController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"200":{"description":"Task updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestError"}}}}},"tags":["Tasks"],"security":[{"Bearer":[]}]},"delete":{"operationId":"TasksController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Task deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestError"}}}}},"tags":["Tasks"],"security":[{"Bearer":[]}]}}},"info":{"title":"NestJS API","description":"An API scaffolded with NestJS","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"Bearer":{"scheme":"bearer","bearerFormat":"Token","type":"http"}},"schemas":{"CreateTaskDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}},"required":["title"]},"Task":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","status","createdAt","updatedAt"]},"BadRequestError":{"type":"object","properties":{"status":{"type":"number"},"message":{"type":"string"},"error":{"type":"string"}},"required":["message","error"]},"UpdateTaskDto":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string","examples":["open","pending","closed"]}},"required":["id"]}}}}